pyproject.toml
[project]
name = "therapy-compliance-processor"
version = "2.0.0"
description = "Automated therapy and medical note compliance checking with MDM analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pdfplumber>=0.10.0",
    "pymupdf>=1.24.0",
    "openai>=1.0.0",
    "python-dateutil>=2.8.2",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.11.0",
]

[project.scripts]
check-compliance = "psychotherapy.therapy_compliance_processor:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["psychotherapy"]

[tool.uv]
dev-dependencies = []
